chore: rename routes with pages
